Getting rid of warnings (using clang) and make some changes to the build

+++ agent/bench.sh
@@ -1,11 +1,16 @@
 #!/bin/sh
 
-#Path to dacpo.jar
-DACAPO="/home/erik/Desktop/dacapo-9.12-bach.jar"
+#Path to dacapo jar
+DACAPO="../dacapo/dacapo-9.12-bach.jar"
+
+if [ ! -e $DACAPO ] ; then
+    echo "can't find <$DACAPO>"
+    echo "please download from http://www.dacapobench.org/ and rerun this script"
+    exit 1
+fi
 
 #classpath
-CP="-Xbootclasspath/p:/home/erik/java-alias-agent/agent/lib/asm-all-4.2.jar:/home/erik/java-alias-agent/agent/lib/awio.jar"
-
+CP="-Xbootclasspath/p:./lib/asm-all-4.2.jar:./lib/awio.jar"
 
 java -noverify -cp $DACAPO $CP -agentpath:./lib/testagent.so Harness $1 -s small
